+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - `openModal` and `openWindow` returns the Stage
 - `dialog` builder operates on a StageAwareFieldset so it can close the dialog easier by calling `close()`
 - All JSON extractor functions support vararg keys, will pick the first available (https://github.com/edvin/tornadofx/issues/350)
+- ValidationContext.validate(decorateErrors = false) clears decorators
 
 ## [1.7.7] - 2017-06-15