+++ CHANGELOG.md
@@ -4,10 +4,13 @@
 
 ### Additions
 
+- ViewModel.clearDecorators()
+
 ### Fixed
 
 ### Changes
 
+- Kotlin 1.1.3
 - `openModal` and `openWindow` returns the Stage
 - `dialog` builder operates on a StageAwareFieldset so it can close the dialog easier by calling `close()`
 - All JSON extractor functions support vararg keys, will pick the first available (https://github.com/edvin/tornadofx/issues/350)